backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.7
apex-saas-framework
python-multipart==0.0.6
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

